pyproject.toml
[project]
name = "activity-info-runner"
version = "0.1.0"
description = "A comprehensive platform for executing automated scripts and workflows against ActivityInfo databases."
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "pydantic",
    "python-dotenv",
    "lark",
    "aiocache[redis]>=0.12.3",
    "rich>=14.2.0",
    "fastapi",
    "uvicorn",
    "temporalio",
    "neo4j",
    "aiofiles>=25.1.0",
]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "respx>=0.22.0",
]
